Server, Redis, Users, and Rooms services implemented
Each service is a class which keeps a singleton instance of itself.
To interact with the services, 'require' them in the file access via
class methods. For example:

const Users = require('../server/users.service');

(...some code)

    socket.on('some event', async (argument) => {
        let user = await Users.get(userID);
    });

(...some other code)

The functions are mostly async, whic means they return Promises objects.
Using await, you can wait for the Promise to be fulfilled and the
expected value to be available.

Signed-off-by: Anderson Toshiyuki Sasaki <anderson.sasaki@gmail.com>